requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.3            # via black
astroid==2.3.3            # via pylint
attrs==19.3.0             # via black
beautifulsoup4==4.8.2     # via -r requirements.in
black==19.10b0            # via -r requirements.in
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.1.1              # via black
idna==2.9                 # via requests
isort==4.3.21             # via -r requirements.in, pylint
lazy-object-proxy==1.4.3  # via astroid
mccabe==0.6.1             # via pylint
mypy-extensions==0.4.3    # via mypy
mypy==0.770               # via -r requirements.in
pathspec==0.7.0           # via black
pylint==2.4.4             # via -r requirements.in
pyyaml==5.3.1             # via -r requirements.in
regex==2020.2.20          # via black
requests==2.23.0          # via -r requirements.in
six==1.14.0               # via astroid
soupsieve==2.0            # via beautifulsoup4
toml==0.10.0              # via black
typed-ast==1.4.1          # via astroid, black, mypy
typing-extensions==3.7.4.1  # via mypy
urllib3==1.25.8           # via requests
wrapt==1.11.2             # via astroid
